+++ src/test/java/com/ejoongseok/wmslive/inbound/feature/ConfirmInboundTest.java
@@ -12,7 +12,6 @@ import org.mockito.Mockito;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ConfirmInboundTest {
-    private final InboundFixture inboundFixture = new InboundFixture();
     private ConfirmInbound confirmInbound;
     private InboundRepository inboundRepository;
 
@@ -38,18 +37,4 @@ class ConfirmInboundTest {
         assertThat(inbound.getStatus()).isEqualTo(InboundStatus.CONFIRMED);
     }
 
-    private class ConfirmInbound {
-
-        private final InboundRepository inboundRepository;
-
-        private ConfirmInbound(final InboundRepository inboundRepository) {
-            this.inboundRepository = inboundRepository;
-        }
-
-        public void request(final Long inboundNo) {
-            final Inbound inbound = inboundRepository.getBy(inboundNo);
-
-            inbound.confirmed();
-        }
-    }
 }
